% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-JavaTests.R
\name{ad.test}
\alias{ad.test}
\title{Anderson-Darling test for goodness of fit}
\usage{
ad.test(sequence, distribution = "NORMAL", mean = NULL, sd = NULL,
  start = NULL, end = NULL, freedom = NULL, k = NULL,
  lambda = NULL, scale = NULL, s = NULL)
}
\arguments{
\item{sequence}{Sequence of data}

\item{distribution}{Distribution name to perform test}

\item{mean}{Distribution mean - Use if mean is a parameter of the distribution}

\item{sd}{Distribution sd - Use if mean is a parameter of the distribution}

\item{start}{Distribution start - Use if mean is a parameter of the distribution}

\item{end}{Distribution end - Use if mean is a parameter of the distribution}

\item{freedom}{Distribution freedom - Use if mean is a parameter of the distribution}

\item{k}{Distribution k - Use if mean is a parameter of the distribution}

\item{lambda}{Distribution lambda - Use if mean is a parameter of the distribution}

\item{scale}{Distribution scale - Use if mean is a parameter of the distribution}

\item{s}{Distribution s - Use if mean is a parameter of the distribution}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Anderson-Darling test
}
