% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AdjustFormatTable}
\alias{AdjustFormatTable}
\title{Adjust Format Table}
\usage{
AdjustFormatTable(table, rownames = NULL, colnames = NULL, ...)
}
\arguments{
\item{table}{Table with p-values}

\item{rownames}{Rownames for the table}

\item{colnames}{Colnames for the table}

\item{...}{Extra arguments for xtable}
}
\value{
Latex code with the formatted table.
}
\description{
Auxiliar method for printing a p-value table with bold font in p-values less than 0.05
}
