% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionTables.R
\name{attParse}
\alias{attParse}
\title{Parse XML attributes to a data frame}
\usage{
attParse(x)
}
\arguments{
\item{x}{Row}
}
\value{
Row with values and attribute names
}
\description{
Parse XML Attributes of a row
}

