% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian-IDP.R
\name{bayesian.imprecise}
\alias{bayesian.imprecise}
\title{Bayesian Imprecise Dirichlet Prior}
\usage{
bayesian.imprecise(x, y, s = sqrt(2) - 1, c = 0.5, mc.samples = 10000)
}
\arguments{
\item{x}{First vector of observations}

\item{y}{Second vector of observations}

\item{s}{Prior strength. Default sqrt(2) - 1}

\item{c}{Probability needed of one algorithm outperforming the other. Default 0.5.}

\item{mc.samples}{Number of Monte Carlo samples.}
}
\value{
List with posterior distribution and probability of exceeding c.
}
\description{
Bayesian Test for two samples based on the Imprecise Dirichlet Theory.
}
