% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian-MultipleMeasurestTest.R
\name{bayesian.MultipleConditions}
\alias{bayesian.MultipleConditions}
\title{Bayesian Test for Multiple Performance Measures Comparison}
\usage{
bayesian.MultipleConditions(x, y, n.samples = 10000, prior = rep(2^-ncol(x),
  ncol(x)))
}
\arguments{
\item{x}{Performance matrix of first algorithm}

\item{y}{Performance matrix of second algorithm}

\item{n.samples}{Size of Monte Carlo sampling}

\item{prior}{Prior distribution parameters}
}
\value{
List with the probabilites of each possible dominance
configuration.
}
\description{
Performs the GLRT for the joint analysis of multiple
performance measures.
}
