% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian-CorrelatedT.R
\name{bayesianCorrelatedT.test}
\alias{bayesianCorrelatedT.test}
\title{Bayesian Correlayed t-test}
\usage{
bayesianCorrelatedT.test(x, y = NULL, rho = 1/length(x),
  rope.min = -0.01, rope.max = 0.01)
}
\arguments{
\item{x}{First vector of observations}

\item{y}{Second vector of observations}

\item{rho}{Expectated correlation.}

\item{rope.min}{Lower limit of the rope considered}

\item{rope.max}{Upper limit of the rope considered}
}
\value{
List with probabilities for each region and a sample of
posterior distribution.
}
\description{
This function performs the Bayesian correlated t-test
}
\examples{
bayesianCorrelatedT.test(results.rf[1, ], results.knn[1, ])
bayesianCorrelatedT.test(results.rf[5, ], results.knn[5, ], rope.min=-0.05, rope.max = 0.05)
}
