% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian-Friedman.R
\name{bayesianFriedman.test}
\alias{bayesianFriedman.test}
\title{Bayesian Friedman Test}
\usage{
bayesianFriedman.test(dataset, s = 1, gamma = 0.05, ties = "average",
  n.samples = 10000, imprecise = FALSE)
}
\arguments{
\item{dataset}{Matrix observations with treatments by columns and observations by rows}

\item{s}{Prior strength. Default 1}

\item{gamma}{Posterior probability level. Default 0.05}

\item{ties}{Method for ties. Default "average"}

\item{n.samples}{Number of samples. Default 10000}

\item{imprecise}{Imprecise method of calculation. Default FALSE}
}
\value{
List with the probabilites of each possible dominance
configuration.
}
\description{
Performs the Bayesian Friedman Test
}
