% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{bayesianSignedRank.test}
\alias{bayesianSignedRank.test}
\title{Bayesian signed rank test}
\usage{
bayesianSignedRank.test(x, y = NULL, s = 0.5, z_0 = 0, rope.min = -0.01,
  rope.max = 0.01, weights = c(s, rep(1, length(x))), samples = 30000)
}
\arguments{
\item{x}{First vector of observations}

\item{y}{Second vector of observations}

\item{s}{Prior pseudo-observation probabilitie}

\item{z_0}{Prior pseudo-observation}

\item{rope.min}{Inferior limit of the rope considered}

\item{rope.max}{Superior limit of the rope considered}

\item{weights}{A priori weights}

\item{samples}{Number of samples of the distribution}
}
\description{
This function performs the bayesian signed rank test
}

