% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{binomialSign.test}
\alias{binomialSign.test}
\title{Sign test for one sample}
\usage{
binomialSign.test(matrix)
}
\arguments{
\item{matrix}{Sequence of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Sign test
}
