% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-EqualityTests.R
\name{cd.test}
\alias{cd.test}
\title{Chakraborti and Desu test for equality}
\usage{
cd.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Chakraborti and Desu test
}
\examples{
cd.test(results)
}
