% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cec17.R
\docType{data}
\name{cec17.final}
\alias{cec17.final}
\title{Final mean results of CEC'17 algorithms}
\format{A data frame with 1276 rows and 4 variables:
\describe{
\item{Algorithm}{Name of the algorithm. Value in \{DES, DYYPO, IDEbestNsize, jSO, LSHADE_SPACMA, MM_OED, PPSO, RB-IPOP-CMA-ES, TLBO-FL, LSHADE-CNEPSIN, EBO-CMAR\}}
\item{Benchmark}{Benchmark function, 1,3-30}
\item{Dimension}{Dimension of the benchmark function: \{10,30,50,100\}}
\item{Result}{Result at 100\\% of iterations.}
}}
\usage{
cec17.final
}
\description{
A dataset containing the final results of algorithms
presented in the CEC'17 competition.
}
\keyword{datasets}
