% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-BivariateTests.R, R/utils.R
\name{checkBivariateConditions}
\alias{checkBivariateConditions}
\alias{checkBivariateConditions}
\title{Checks Conditions}
\usage{
checkBivariateConditions(matrix)

checkBivariateConditions(matrix)
}
\arguments{
\item{matrix}{Matrix of data}

\item{matrix}{Matrix of data}
}
\description{
Checks conditions for bivariate non parametric tests

Checks conditions for bivariate non parametric tests
}
