% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-CountDataTests.R
\name{checkCountDataConditions}
\alias{checkCountDataConditions}
\title{Checks Conditions}
\usage{
checkCountDataConditions(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\description{
Checks conditions for count data non parametric tests
}

