% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT-MultipleMeasuresTest.R
\name{checkMultipleMeasuresConditions}
\alias{checkMultipleMeasuresConditions}
\title{Check Multiple Measures Conditions}
\usage{
checkMultipleMeasuresConditions(x, y)
}
\arguments{
\item{x}{Performance matrix of first algorithm}

\item{y}{Performance matrix of second algorithm}
}
\description{
Check the safety conditions for Multiple Measures Tests
}
