% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{chiSquare.test}
\alias{chiSquare.test}
\title{Chi square test for goodness of fit}
\usage{
chiSquare.test(matrix, n, p = NULL)
}
\arguments{
\item{matrix}{Matrix of data}

\item{n}{N parameter for Uniform distribution}

\item{p}{P parameter for uniform distribution}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Chi square test
}

