% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionTables.R
\name{computeAproximatedProbability}
\alias{computeAproximatedProbability}
\title{Compute exact probability of distribution}
\usage{
computeAproximatedProbability(table, n, T)
}
\arguments{
\item{table}{Distribution table}

\item{n}{Size of the exact distribution}

\item{T}{Value of the statistic}
}
\value{
p-value computed
}
\description{
Function to get the exact probability given a distribution table
}

