% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-EqualityTests.R
\name{computeCDExactProbability}
\alias{computeCDExactProbability}
\title{Chakraborti and Desu Exact Probability}
\usage{
computeCDExactProbability(W, N, populations, length)
}
\arguments{
\item{W}{W statistic}

\item{N}{Number of elements}

\item{populations}{Number of populations}

\item{length}{Number values per sample}
}
\value{
p-value computed
}
\description{
This function compute the exact probability for Chakraborti and Desu test
}

