% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-CountDataTests.R
\name{computeFisherRightExactProbability}
\alias{computeFisherRightExactProbability}
\title{Compute Fisher Right Exact Probability}
\usage{
computeFisherRightExactProbability(N, n1, n2, Y, n00)
}
\arguments{
\item{N}{Number of samples}

\item{n1}{Sum of the samples of the first row}

\item{n2}{sum of the samples of the second row}

\item{Y}{sum of the samples of the first column}

\item{n00}{samples in first column, first row}
}
\value{
p-value computed
}
\description{
Computes exact left tail of Fisher distribution
}

