% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-TwoSamplesTests.R
\name{computeKolmogorovAsymptoticProbability}
\alias{computeKolmogorovAsymptoticProbability}
\title{Asymptotic p values for Kolmogorov-Smirnov test for compare two samples}
\usage{
computeKolmogorovAsymptoticProbability(n, Dn)
}
\arguments{
\item{n}{Size of distribution}

\item{Dn}{Statistic}
}
\value{
computed pvalue
}
\description{
This function compute the asymptotic pvalue for the Kolmogorov-Smirnov test for two samples
}

