% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-TwoSamplesTests.R
\name{computeKolmogorovExactProbability}
\alias{computeKolmogorovExactProbability}
\title{Exact p values for Kolmogorov-Smirnov test for compare two samples}
\usage{
computeKolmogorovExactProbability(n, Dn)
}
\arguments{
\item{n}{Size of distribution}

\item{Dn}{Statistic}
}
\value{
computed pvalue
}
\description{
This function compute the exact pvalue for the Kolmogorov-Smirnov test for two samples
}

