% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-RandomnessTests.R
\name{computeNumberOfRunsAsymptoticProbability}
\alias{computeNumberOfRunsAsymptoticProbability}
\title{Asymptotic values}
\usage{
computeNumberOfRunsAsymptoticProbability(n, R)

computeNumberOfRunsAsymptoticProbability(n, R)
}
\arguments{
\item{n}{Number of elements}

\item{R}{Runs up down statistic}

\item{a}{Number of elements of first class}

\item{b}{Number of elements of second class}

\item{runs}{Number of runs}
}
\value{
pvalues computed

Asymptotic pvalues computed
}
\description{
Computes asymptotic p-values of the Total number of runs distribution

Computes asymptotic p-values of the Total number of runs up and down distribution
}

