% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-RandomnessTests.R
\name{computeNumberOfRunsRightTailProbability}
\alias{computeNumberOfRunsRightTailProbability}
\title{Exact Left tail probability fo number of runs distribution}
\usage{
computeNumberOfRunsRightTailProbability(a, b, runs)
}
\arguments{
\item{a}{Number of elements of first class}

\item{b}{Number of elements of second class}

\item{runs}{Number of runs}
}
\value{
pvalue computed
}
\description{
Computes left tail p-value of the Total number of runs distribution
}

