% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-RandomnessTests.R
\name{computeNumberOfUpDownRunsAsymptoticProbability}
\alias{computeNumberOfUpDownRunsAsymptoticProbability}
\title{Asymptotic values of runs up and down test}
\usage{
computeNumberOfUpDownRunsAsymptoticProbability(n, R)
}
\arguments{
\item{n}{Number of elements}

\item{R}{Runs up down statistic}
}
\value{
Asymptotic pvalues computed
}
\description{
Computes asymptotic p-values of the Total number of runs up and down distribution
}
