% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-MultipleTests.R
\name{computePageAsymptoticProbability}
\alias{computePageAsymptoticProbability}
\title{Compute asymptotic probability of distribution}
\usage{
computePageAsymptoticProbability(N, k, L)
}
\arguments{
\item{N}{number of columns}

\item{k}{number of rows}

\item{L}{Page statistic}
}
\value{
Exact p-value computed
}
\description{
Function to get the asymptotic probability given a distribution table
}

