% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-MultipleTests.R
\name{computePageExactProbability}
\alias{computePageExactProbability}
\title{Compute exact probability of Page distribution}
\usage{
computePageExactProbability(N, k, L)
}
\arguments{
\item{N}{number of columns}

\item{k}{number of rows}

\item{L}{Page statistic}
}
\value{
Exact p-value computed
}
\description{
Function to get the exact probability of Page distribution
}

