% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-RandomnessTests.R
\name{computeRunsUpDownExactProbability}
\alias{computeRunsUpDownExactProbability}
\title{Exact tail probability fo number of runs distribution}
\usage{
computeRunsUpDownExactProbability(n, NM, RVN)

computeRunsUpDownExactProbability(n, NM, RVN)
}
\arguments{
\item{n}{Number of elements}

\item{NM}{NM statistic}

\item{RVN}{RVN statistic}

\item{R}{Runs up down statistic}

\item{n}{Number of elements}
}
\value{
pvalue computed

pvalue computed
}
\description{
Computes exact p-value of the Runs up down distribution

Computes exact p-value of the Von Neumann distribution
}

