% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-RandomnessTests.R
\name{computeTotalNumberOfRunsAsymptoticProbability}
\alias{computeTotalNumberOfRunsAsymptoticProbability}
\title{Asymptotic values}
\usage{
computeTotalNumberOfRunsAsymptoticProbability(n1, n2, runs)
}
\arguments{
\item{n1}{Number of elements of first class}

\item{n2}{Number of elements of second class}

\item{runs}{Number of runs}
}
\value{
pvalues computed
}
\description{
Computes asymptotic p-values of the Total number of runs distribution
}
