% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-RandomnessTests.R
\name{computeVonNewmannExactProbability}
\alias{computeVonNewmannExactProbability}
\title{Exact tail probability for Von Neumann distribution}
\usage{
computeVonNewmannExactProbability(n, NM, RVN)
}
\arguments{
\item{n}{Number of elements}

\item{NM}{NM statistic}

\item{RVN}{RVN statistic}
}
\value{
pvalue computed
}
\description{
Computes exact p-value of the Von Neumann distribution
}

