% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-OneSampleTests.R
\name{computeWilcoxonAsymptoticProbability}
\alias{computeWilcoxonAsymptoticProbability}
\title{Compute asymptotic probability of Wilcoxon distribution}
\usage{
computeWilcoxonAsymptoticProbability(n, R, ties)
}
\arguments{
\item{n}{Size of distribution}

\item{R}{statistic}

\item{ties}{Ties}
}
\value{
Exact p-value computed
}
\description{
Function to get asymptotic probability of Wilcoxon distribution
}

