% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-OneSampleTests.R
\name{computeWilcoxonExactProbability}
\alias{computeWilcoxonExactProbability}
\title{Compute exact probability of Wilcoxon distribution}
\usage{
computeWilcoxonExactProbability(n, R)
}
\arguments{
\item{n}{Size of distribution}

\item{R}{statistic}
}
\value{
Exact p-value computed
}
\description{
Function to get the exact probability of Wilcoxon distribution
}

