% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-LocationTests.R
\name{computeWilcoxonRankLeftProbability}
\alias{computeWilcoxonRankLeftProbability}
\title{Wilcoxon exact left p-value}
\usage{
computeWilcoxonRankLeftProbability(n, m, R)
}
\arguments{
\item{n}{n parameter}

\item{m}{m parameter}

\item{R}{rank associated}
}
\value{
pvalue associated
}
\description{
Compute left p-value for Wilcoxon rank sum test
}

