% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-LocationTests.R
\name{computeWilcoxonRankPValues}
\alias{computeWilcoxonRankPValues}
\title{Wilcoxon Rank p-values}
\usage{
computeWilcoxonRankPValues(combined, n1, n2, WRank)
}
\arguments{
\item{combined}{Vector of combined samples}

\item{n1}{Size of first sample}

\item{n2}{Size of second sample}

\item{WRank}{Statistic}
}
\value{
pvalues
}
\description{
Compute p-values for Wilcoxon
}

