% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-LocationTests.R
\name{computeWilcoxonRankRightProbability}
\alias{computeWilcoxonRankRightProbability}
\title{Wilcoxon exact right p-value}
\usage{
computeWilcoxonRankRightProbability(n, m, R)
}
\arguments{
\item{n}{n parameter}

\item{m}{m parameter}

\item{R}{rank associated}
}
\value{
pvalue associated
}
\description{
Compute right p-value for Wilcoxon rank sum test
}

