% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-JavaTests.R
\name{concordanceCoeff.test}
\alias{concordanceCoeff.test}
\title{Concordance Coefficient test for multiple comparisons}
\usage{
concordanceCoeff.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Concordance Coefficient test
}
