% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{confidenceQuantile.test}
\alias{confidenceQuantile.test}
\title{Confidence Quantile for one sample}
\usage{
confidenceQuantile.test(n, p, q)
}
\value{
Report of the test in a string
}
\description{
This function performs the confidence quantile test
}

