% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{correlatedBayesianT.test}
\alias{correlatedBayesianT.test}
\title{Bayesian signed rank test}
\usage{
correlatedBayesianT.test(x, y = NULL, rho = 1/length(x), rope.min = -0.01,
  rope.max = 0.01)
}
\arguments{
\item{x}{First vector of observations}

\item{y}{Second vector of observations}

\item{rho}{Expectated correlation.}

\item{rope.min}{Lower limit of the rope considered}

\item{rope.max}{Upper limit of the rope considered}
}
\description{
This function performs the bayesian signed rank test
}

