% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create.permutations}
\alias{create.permutations}
\title{Create Permutations}
\usage{
create.permutations(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Data frame with all the different permutations.
}
\description{
Auxiliar method for create permutations with the different possible results for a heaviside vector
}
