% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-BivariateTests.R
\name{danielTrend.test}
\alias{danielTrend.test}
\title{Daniel Trend test for bivariated samples}
\usage{
danielTrend.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Daniel Trend test
}
\examples{
x <- 1:10
m <- matrix(c(x, 2*x+rnorm(length(x))), ncol = 2)
danielTrend.test(m)
}

