% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{dataTable}
\alias{dataTable}
\title{Java DataTable object}
\usage{
dataTable(matrix)
}
\arguments{
\item{matrix}{Matrix to create the object}
}
\value{
Java DataTable object
}
\description{
Returns a DataTable object
}

