% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-JavaTests.R
\name{davidBarton.test}
\alias{davidBarton.test}
\title{David Barton test for scale}
\usage{
davidBarton.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the David Barton test
}
