% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{Automatic execution function on load of package}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Name of the library}

\item{pkgname}{Name of the package}
}
\description{
This function executes when package is loades
}
