% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{doTest}
\alias{doTest}
\title{Carry out a test introducing name and data}
\usage{
doTest(test.name, ...)
}
\arguments{
\item{test.name}{Name of the test.}
}
\value{
The function returns the report of the test in a string
}
\description{
Function to make a non-parametric test
}

