% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionTables.R
\name{doubleTailProbability}
\alias{doubleTailProbability}
\title{Double tail probability}
\usage{
doubleTailProbability(l, r)
}
\arguments{
\item{l}{Left tail probability}

\item{r}{Right tail probability}
}
\value{
Double tail probability
}
\description{
Get double tail probability
}

