% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{extendedMedian.test}
\alias{extendedMedian.test}
\title{Extended Median test for equality}
\usage{
extendedMedian.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the extended median test
}
