% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{fillSequence}
\alias{fillSequence}
\title{Fill a sequence with an array}
\usage{
fillSequence(sequece, array)
}
\arguments{
\item{array}{Array with the data to fill the sequence up}
}
\value{
Java sequece object
}
\description{
Function to fill a sequece
}

