% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{freundAnsariBradley.test}
\alias{freundAnsariBradley.test}
\title{Freund Ansari Bradley test for scale}
\usage{
freundAnsariBradley.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Freund Ansari Bradley test
}

