% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{friedman.test}
\alias{friedman.test}
\title{Friedman test for multiple comparisons}
\usage{
friedman.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Friedman test
}

