% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionTables.R
\name{getCumulativeProbabilityFunction}
\alias{getCumulativeProbabilityFunction}
\title{Get cumulative probability function}
\usage{
getCumulativeProbabilityFunction(distribution, ...)
}
\arguments{
\item{distribution}{Distribution name}

\item{...}{Parameters for the distribution}
}
\value{
Cumulative probability function
}
\description{
Get cumulative probability function according to distribution
}
