% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getdata}
\alias{getdata}
\title{Get data from data folderFunction to get a data file and assign it to a variable}
\usage{
getdata(...)
}
\arguments{
\item{...}{Parameters to data function}
}
\value{
Data to assign it to a variable
}
\description{
Function to get a data file and assign it to a variable
}

