% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{getDiff}
\alias{getDiff}
\title{Auxiliar function to get the difference between observations and}
\usage{
getDiff(x, y)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\description{
This function returns the difference between two vectors if they have the same length. If the second vector is missing, the first one is returned
}

