% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionTables.R
\name{getFromDistributionTable}
\alias{getFromDistributionTable}
\title{Get value from distribution table}
\usage{
getFromDistributionTable(table, n.value, T.value, epsilon = 0.002)
}
\arguments{
\item{table}{Distribution table}

\item{epsilon}{Threshold for compare the statistic}

\item{n}{Size of the exact distribution}

\item{T}{Value of the statistic}
}
\value{
p-value computed
}
\description{
Function to get the exact probability given a distribution table
}

