% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{heaviside}
\alias{heaviside}
\title{Heaviside step function}
\usage{
heaviside(x, a = 0)
}
\arguments{
\item{x}{Object to be evaluated}

\item{a}{Frontier value}
}
\description{
This function implements Heaviside step function
}

