% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{htest2Tex.list}
\alias{htest2Tex.list}
\title{Test object to table in LaTeX format}
\usage{
\method{htest2Tex}{list}(test)
}
\arguments{
\item{test}{Test object with pvalue(s), test name and statistic(s)}
}
\value{
This method prints the necessary code for include a table
with the information provided by the test.
}
\description{
Transform a test object to table in LaTeX format
}
\examples{
htest2Tex(cd.test(results))
}
