% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-ImanDavenport.R
\name{imanDavenport.test}
\alias{imanDavenport.test}
\title{Iman-Davenport test}
\usage{
imanDavenport.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Iman-Davenport test
}
\examples{
imanDavenport.test(results)
}
