% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-JavaTests.R
\name{jt.test}
\alias{jt.test}
\title{Jonckheere and Terpstra test for equality}
\usage{
jt.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Jonckheere and Terpstra test
}
