% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-BivariateTests.R
\name{kendall.test}
\alias{kendall.test}
\title{Kendall test for bivariated samples}
\usage{
kendall.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Kendall test
}
\examples{
x <- 1:10
m <- matrix(c(x, 2*x+rnorm(length(x))), ncol = 2)
kendall.test(m)
}

