% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{KendallExactTable}
\alias{KendallExactTable}
\title{Table for Kendall test}
\format{An object of class \code{data.frame} with 88 rows and 3 columns.}
\usage{
KendallExactTable
}
\description{
Table for Kendall test
}
\keyword{datasets}

