% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{KolmogorovTable}
\alias{KolmogorovTable}
\title{Table for KolmogorovTable}
\format{An object of class \code{matrix} with 40 rows and 5 columns.}
\usage{
KolmogorovTable
}
\description{
Table for KolmogorovTable
}
\keyword{datasets}

