% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{KolmogorovTwoSample}
\alias{KolmogorovTwoSample}
\title{Table for KolmogorovTwoSample}
\format{An object of class \code{data.frame} with 196 rows and 4 columns.}
\usage{
KolmogorovTwoSample
}
\description{
Table for KolmogorovTwoSample
}
\keyword{datasets}

