% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{kruskalWallis.test}
\alias{kruskalWallis.test}
\title{Kruskal-Wallis est for equality}
\usage{
kruskalWallis.test(matrix, print.multiple = F)
}
\arguments{
\item{matrix}{Matrix of data}

\item{print.multiple}{If True, prints}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Kruskal-Wallis test
}
