% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-GoodnessTests.R
\name{ks.test}
\alias{ks.test}
\title{Kolmogorov-Smirnov test for goodness of fit}
\usage{
ks.test(sequence, distribution = "NORMAL", ...)
}
\arguments{
\item{sequence}{Secuence of data}

\item{distribution}{Distribution name to perform test}

\item{...}{Parameters for the distribution}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Kolmogorov-Smirnov test
}

