% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-TwoSamplesTests.R
\name{ksTwoSamples.test}
\alias{ksTwoSamples.test}
\title{Kolmogorov-Smirnov test for compare two samples}
\usage{
ksTwoSamples.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Kolmogorov-Smirnov test for two samples
}

