% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{lilliefors.test}
\alias{lilliefors.test}
\title{Lilliefors test for goodness of fit}
\usage{
lilliefors.test(sequence)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Lilliefors test
}

