% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locate.max}
\alias{locate.max}
\title{Location of the maximum(s) in a vector}
\usage{
locate.max(x)
}
\arguments{
\item{x}{Vector where find the maximum(s)}
}
\description{
Returns a vector with an 1 in the maximum position. If there are more than one then 1/num. maximums is situated in each position
}

