% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{make.htest}
\alias{make.htest}
\title{Make a htest object}
\usage{
make.htest(...)
}
\arguments{
\item{...}{Values to form the htest object}
}
\value{
htest object
}
\description{
This function takes the arguments, makes a list and assign the list the htest class
}
