% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-CountDataTests.R
\name{mcNemar.test}
\alias{mcNemar.test}
\title{McNemar test for counts of data}
\usage{
mcNemar.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the McNemar test
}
\examples{
mcNemar.test(matrix(c(10,4,2,7), ncol=2))
}

