% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT-MultipleMeasuresTest.R
\name{multipleMeasuresGLRT}
\alias{multipleMeasuresGLRT}
\title{Generalized Likelihood Ratio Test for Multiple Performance
Measures}
\usage{
multipleMeasuresGLRT(x, y)
}
\arguments{
\item{x}{Performance matrix of first algorithm}

\item{y}{Performance matrix of second algorithm}
}
\value{
List with the probabilites of each possible dominance
configuration.
}
\description{
Performs the GLRT for the joint analysis of multiple
performance measures.
}
