% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{NMRanksLeft}
\alias{NMRanksLeft}
\title{Table for NMRanksLeft}
\format{An object of class \code{data.frame} with 49 rows and 3 columns.}
\usage{
NMRanksLeft
}
\description{
Table for NMRanksLeft
}
\keyword{datasets}

