% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{NMRanksRight}
\alias{NMRanksRight}
\title{Right distribution table for exact Von Neumann test}
\format{An object of class \code{data.frame} with 54 rows and 3 columns.}
\usage{
NMRanksRight
}
\description{
Right distribution table for exact Von Neumann test
}
\keyword{datasets}
