% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{normalScores.test}
\alias{normalScores.test}
\title{Normal scores test for location}
\usage{
normalScores.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the normal scores test
}

