% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-ApproximateConfidenceInterval.R
\name{NP.ApproximateConfidenceInterval}
\alias{NP.ApproximateConfidenceInterval}
\title{Approximate Non-Parametric Confidence Interval}
\usage{
NP.ApproximateConfidenceInterval(x, y, alpha = 0.05, paired = FALSE)
}
\arguments{
\item{x}{First algorithm's results}

\item{y}{Second algorithm's results}

\item{alpha}{100(1-alpha)\% level of significance}

\item{paired}{Boolean that indicates if the samples are paired}
}
\value{
NP confidence interval
}
\description{
Approximate and Non-Parametric approach to the computation of the Confidence Interval for the comparison of two algorihms
}
