% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-RandomnessTests.R
\name{numberRunsUpDown.test}
\alias{numberRunsUpDown.test}
\title{Number of runs up and down median test for randomness}
\usage{
numberRunsUpDown.test(sequence)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Number of runs up and down  test
}
\examples{
 numberRunsUpDown.test(c(23,99,89,36,1,64,14,28,2,81))
}

