% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-RandomnessTests.R
\name{numberRunsUpDownMedian.test}
\alias{numberRunsUpDownMedian.test}
\title{Number of runs up and down median test for randomness}
\usage{
numberRunsUpDownMedian.test(sequence)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Number of runs up and down median test
}
\examples{
 numberRunsUpDownMedian.test(c(23,99,89,36,1,64,14,28,2,81))
}

