% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{numericSequence}
\alias{numericSequence}
\title{Create a numeric sequence}
\usage{
numericSequence(array)
}
\arguments{
\item{array}{Array with the data to fill the sequence up}
}
\value{
Java String sequece object
}
\description{
Function to create a numeric sequence
}

