% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{occurencesDominanceConfiguration}
\alias{occurencesDominanceConfiguration}
\title{Occurences Dominance Configuration}
\usage{
occurencesDominanceConfiguration(x, y)
}
\arguments{
\item{x}{Performance matrix of first algorithm}

\item{y}{Performance matrix of second algorithm}
}
\value{
Occurence count vector
}
\description{
Count the occurences for each possible dominance
configuration
}
