% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{orderedEq.test}
\alias{orderedEq.test}
\title{Ordered equality test for count data}
\usage{
orderedEq.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Ordered equality test
}
