% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{PageTable}
\alias{PageTable}
\title{Table for PageTable}
\format{An object of class \code{data.frame} with 198 rows and 4 columns.}
\usage{
PageTable
}
\description{
Table for PageTable
}
\keyword{datasets}

