% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-MultipleTests.R
\name{partialcorrelation.test}
\alias{partialcorrelation.test}
\title{Partial Correlation test for multiple comparisons}
\usage{
partialcorrelation.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Partial correlation test
}

