% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{PartialCorrelationTable}
\alias{PartialCorrelationTable}
\title{Distribution table for Partial Correlation test}
\format{An object of class \code{matrix} with 20 rows and 4 columns.}
\usage{
PartialCorrelationTable
}
\description{
Distribution table for Partial Correlation test
}
\keyword{datasets}
