% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionTables.R
\name{pkolmogorov}
\alias{pkolmogorov}
\title{Kolmogorov probability}
\usage{
pkolmogorov(n, Dn)
}
\arguments{
\item{n}{Size of the population}

\item{Dn}{Kolmogorov statistic}
}
\value{
p-value computed
}
\description{
Computes p-value of the Kolmogorov distribution
}

