% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-PosteriorDirichlet.R
\name{plot.PosteriorDirichlet}
\alias{plot.PosteriorDirichlet}
\title{Plot posterior Dirichlet distribution}
\usage{
\method{plot}{PosteriorDirichlet}(x, num.points = nrow(x$sample), ...)
}
\arguments{
\item{x}{Simplex points to be projected}

\item{num.points}{Number of points to be shown}

\item{...}{Extra parameters for plot}
}
\description{
Plot of the projection of 3-simplex points of a posterior Dirichlet distribution into a 2D triangle
}
\examples{
bs <- bayesianSign.test(results$random.forest, results$KNN)
plot(bs, num.points = 5000)

bsr <- bayesianSignedRank.test(results$random.forest, results$KNN)
plot(bsr, num.points = 5000)
}
