% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-PosteriorT.R
\name{plot.PosteriorT}
\alias{plot.PosteriorT}
\title{Plot of posterior t distribution}
\usage{
\method{plot}{PosteriorT}(x, ...)
}
\arguments{
\item{x}{Data of the posterior distribution}

\item{...}{Extra parameters for plot}
}
\description{
This function plots the posterior distribution of the parameter
}
\examples{
cbt <- bayesianCorrelatedT.test(results.rf[1, ], results.knn[1, ])
plot(cbt) + ggplot2::ggtitle("Random Forest vs. KNN \nDataset: abalone")
}
