% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ConfidenceCurve.R
\name{PlotConfidenceCurve}
\alias{PlotConfidenceCurve}
\title{Plot Confidence Curve
Plot confidence intervals for all levels of significance.}
\usage{
PlotConfidenceCurve(x, y, paired = FALSE, alpha = 0.05, m = 10)
}
\arguments{
\item{x}{First algorithm's results}

\item{y}{Second algorithm's results}

\item{paired}{Boolean that indicates if the samples are paired}

\item{alpha}{Significance value}

\item{m}{The number of nested confidence curves to be computed. Default is 100.}
}
\value{
Graphic
}
\description{
Plot Confidence Curve
Plot confidence intervals for all levels of significance.
}
