% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-IDP.R
\name{plotIDP}
\alias{plotIDP}
\title{Plot of posterior distributions of IDP}
\usage{
plotIDP(lower.dist, upper.dist)
}
\arguments{
\item{lower.dist}{Lower Bound Distribution}

\item{upper.dist}{Upper Bound Distribution}
}
\description{
This function plots the posterior distribution of the parameter
}
\examples{
idp <- bayesian.imprecise(results.rf[1, ], results.knn[1, ])
plotIDP(idp$post.dist.lower, idp$post.dist.upper)
}
