% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPosterior}
\alias{plotPosterior}
\title{Plot of posterior distribution}
\usage{
plotPosterior(data, names, dataset)
}
\arguments{
\item{data}{Data for plotting}

\item{names}{Names of the algorithms}

\item{dataset}{Names of the dataset}
}
\description{
This function plots the posterior distribution of the parameter
}
\examples{
cbt <- correlatedBayesianT.test(results.rf[1, ], results.knn[1, ])
plotPosterior(cbt, names = c("Random Forest", "KNN"), dataset = "abalone")
}
