% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSimplex}
\alias{plotSimplex}
\title{Projection of 3-simplex points}
\usage{
plotSimplex(points, num.points = nrow(points))
}
\arguments{
\item{points}{Simplex points to be projected}

\item{num.points}{Number of points to be shown}
}
\description{
This function projects 3-simplex points to a 2D triangle
}
\examples{
bs <- bayesianSign.test(results$random.forest, results$KNN)
plotSimplex(bs$sample, num.points = 5000)

bsr <- bayesianSignedRank.test(results$random.forest, results$KNN)
plotSimplex(bsr$sample, num.points = 5000)
}
