% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{populationQuantile.test}
\alias{populationQuantile.test}
\title{Population quantile for one sample}
\usage{
populationQuantile.test(sequence, quantile, value)
}
\arguments{
\item{sequence}{Parameter of the test}

\item{quantile}{Parameter of the test}

\item{value}{Parameter of the test}
}
\value{
Report of the test in a string
}
\description{
This function performs the confidence quantile test
}

