% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{RanksVonNeumann}
\alias{RanksVonNeumann}
\title{Distribution table for Von Neumann test}
\format{An object of class \code{matrix} with 41 rows and 5 columns.}
\usage{
RanksVonNeumann
}
\description{
Distribution table for Von Neumann test
}
\keyword{datasets}
