% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{data}
\name{results.knn}
\alias{results.knn}
\title{Dataset with CV results of KNN algorithm in different datasets}
\format{An data frame with 29 rows and 5 columns}
\usage{
results.knn
}
\description{
Results of KNN algorithm in different datasets
}
\keyword{datasets}

