% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{data}
\name{results.lr}
\alias{results.lr}
\title{Dataset with CV results of logistic regression algorithm in different datasets}
\format{An data frame with 29 rows and 5 columns}
\usage{
results.lr
}
\description{
Results of logistic regression algorithm in different datasets
}
\keyword{datasets}
