% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{data}
\name{results.nb}
\alias{results.nb}
\title{Dataset with CV results of Naive Bayes algorithm in different datasets}
\format{An data frame with 29 rows and 5 columns}
\usage{
results.nb
}
\description{
Results of Naive Bayes algorithm in different datasets
}
\keyword{datasets}
