% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{data}
\name{results.nnet}
\alias{results.nnet}
\title{Dataset with CV results of neural network algorithm in different datasets}
\format{An data frame with 29 rows and 5 columns}
\usage{
results.nnet
}
\description{
Results of neural network algorithm in different datasets
}
\keyword{datasets}
