% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{data}
\name{results}
\alias{results}
\title{Dataset with CV mean results of some algorithms in different datasets}
\format{An data frame with 29 rows and 5 columns}
\usage{
results
}
\description{
Results of some algotithms in different datasets
}
\keyword{datasets}

