% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-results.R
\docType{data}
\name{results.rf}
\alias{results.rf}
\title{Dataset with CV results of Random Forest algorithm in different datasets}
\format{An data frame with 29 rows and 5 columns}
\usage{
results.rf
}
\description{
Results of Random Forest algorithm in different datasets
}
\keyword{datasets}
