% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-JavaTests.R
\name{runTest}
\alias{runTest}
\title{Execution of a StatisticalTest object}
\usage{
runTest(java.test.object)
}
\arguments{
\item{java.test.object}{A StatisticalTest Java object}
}
\value{
The function returns the report of the test in a string
}
\description{
Wrapper to run a StatisticalTest object.
}
