% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{sign.test}
\alias{sign.test}
\title{Sign test for one sample}
\usage{
\method{sign}{test}(sequence)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Sign test
}

