% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{SpearmanExactTable}
\alias{SpearmanExactTable}
\title{Exact distribution table for Spearman test}
\format{An object of class \code{data.frame} with 253 rows and 3 columns.}
\usage{
SpearmanExactTable
}
\description{
Exact distribution table for Spearman test
}
\keyword{datasets}
