% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{stringSequence}
\alias{stringSequence}
\title{Create a string sequence}
\usage{
stringSequence(array)
}
\arguments{
\item{array}{Array with the data to fill the sequence up}
}
\value{
Java String sequece object
}
\description{
Function to create a string sequence
}

