% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-RandomnessTests.R
\name{vonNeumann.test}
\alias{vonNeumann.test}
\title{Von Neumann test for randomness}
\usage{
vonNeumann.test(sequence)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Von Neumann
}
\examples{
 vonNeumann.test(c(23,99,89,36,1,64,14,28,2,81))
}
