% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-TwoSamplesTests.R
\name{waldWolfowitz.test}
\alias{waldWolfowitz.test}
\title{Wald-Wolfowitz test for compare two samples}
\usage{
waldWolfowitz.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Wald-Wolfowitz test for two samples
}
\examples{
waldWolfowitz.test(results[ ,c(1,2)])
}
