% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-TwoSamplesTests.R
\name{waldWolfowitzSequence}
\alias{waldWolfowitzSequence}
\title{Wald-Wolfowitz char vector construction}
\usage{
waldWolfowitzSequence(sample1, sample2)
}
\arguments{
\item{sample1}{First sample}

\item{sample2}{Second sample}
}
\value{
Sequence to apply Wald-Wolfowitz test
}
\description{
Get sequence for Wald-Wolfowitz test
}

