% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-OneSampleTests.R
\name{wilcoxon.test}
\alias{wilcoxon.test}
\title{Wilcoxon Test for one sample}
\usage{
wilcoxon.test(matrix)
}
\arguments{
\item{matrix}{matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Wilcoxon test
}
\examples{
wilcoxon.test(results[ ,c(1,2)])
}

