% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-LocationTests.R
\name{wilcoxonRankSum.test}
\alias{wilcoxonRankSum.test}
\title{Wilcoxon Rank Sum test for location}
\usage{
wilcoxonRankSum.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Wilcoxon Rank Sum test
}
\examples{
wilcoxonRankSum.test(results[ ,1:2])
}
