% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionTables.R
\name{xmlDistributionToDataFrame}
\alias{xmlDistributionToDataFrame}
\title{Parse XML distribution to a data frame}
\usage{
xmlDistributionToDataFrame(xml.file, name.distribution)
}
\arguments{
\item{xml.file}{XML file with the distribution}

\item{name.distribution}{Name of the distribution}
}
\value{
Distribution data frame
}
\description{
Parse XML distribution to a data frame
}

