% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionTables.R
\name{xmlQuantileDistributionToMatrix}
\alias{xmlQuantileDistributionToMatrix}
\title{Parse XML quantile distribution to a data frame}
\usage{
xmlQuantileDistributionToMatrix(xml.file, pvalues)
}
\arguments{
\item{xml.file}{XML file with the distribution}

\item{pvalues}{p-values of the distribution of the statistic}
}
\value{
Distribution data frame
}
\description{
Parse XML distribution to a data frame
}

