% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{ad.test}
\alias{ad.test}
\title{Anderson-Darling test for goodness of fit}
\usage{
ad.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Anderson-Darling test
}

