% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{ad.test}
\alias{ad.test}
\title{Anderson-Darling test for goodness of fit}
\usage{
ad.test(sequence, distribution = "NORMAL", mean = NULL, sd = NULL,
  start = NULL, end = NULL, freedom = NULL, k = NULL, lambda = NULL,
  scale = NULL, s = NULL)
}
\arguments{
\item{sequence}{Sequence of data}

\item{distribution}{Distribution name to perform test}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Anderson-Darling test
}

