% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian-SignedRank.R
\name{bayesianSignedRank.test}
\alias{bayesianSignedRank.test}
\title{Bayesian Signed-Rank test}
\usage{
bayesianSignedRank.test(
  x,
  y = NULL,
  s = 0.5,
  z_0 = 0,
  rope.min = -0.01,
  rope.max = 0.01,
  weights = NULL,
  mc.samples = 10000
)
}
\arguments{
\item{x}{First vector of observations}

\item{y}{Second vector of observations}

\item{s}{Prior pseudo-observation probabilitie}

\item{z_0}{Prior pseudo-observation}

\item{rope.min}{Inferior limit of the rope considered}

\item{rope.max}{Superior limit of the rope considered}

\item{weights}{A priori weights}

\item{mc.samples}{Number of samples of the distribution}
}
\value{
List with probabilities for each region and a sample of
posterior distribution.
}
\description{
This function performs the Bayesian Signed-Rank test
}
\examples{
bsr <- bayesianSignedRank.test(results$random.forest, results$KNN)
bsr.stronger.prior <- bayesianSignedRank.test(results$random.forest, results$KNN, s=3, z_0 = 0.5)
}
