% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-EqualityTests.R
\name{cd.test}
\alias{cd.test}
\title{Chakraborti and Desu test for equality}
\usage{
cd.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Chakraborti and Desu test
}

