% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{chiSquare.test}
\alias{chiSquare.test}
\title{Chi square test for goodness of fit}
\usage{
chiSquare.test(sequece)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
List of a htest object with pvalues, statistic and method name and a report of the test in a string
}
\description{
This function performs the Chi square test
}

