% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-RandomnessTests.R
\name{computeRunsUpDownExactProbability}
\alias{computeRunsUpDownExactProbability}
\title{Exact tail probability fo number of runs distribution}
\usage{
computeRunsUpDownExactProbability(n, R)
}
\arguments{
\item{n}{Number of elements}

\item{R}{Runs up down statistic}
}
\value{
pvalue computed
}
\description{
Computes exact p-value of the Runs up down distribution
}

