% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{contingency.coeff.test}
\alias{contingency.coeff.test}
\title{Contingency Coefficient test for count data}
\usage{
contingency.coeff.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Contingency Coefficient test
}

