% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{controlMedian.test}
\alias{controlMedian.test}
\title{Control Median test for two samples}
\usage{
controlMedian.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Control Median test
}

