% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{danielTrend.test}
\alias{danielTrend.test}
\title{Daniel Trend test for bivariated samples}
\usage{
danielTrend.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
List of a htest object with pvalues, statistic and method name and a report of the test in a string
}
\description{
This function performs the Daniel Trend test
}

