% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{davidBarton.test}
\alias{davidBarton.test}
\title{David Barton test for scale}
\usage{
davidBarton.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the David Barton test
}

