% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{extendedMedian.test}
\alias{extendedMedian.test}
\title{Extended Median test for equality}
\usage{
extendedMedian.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
List of a htest object with pvalues, statistic and method name and a report of the test in a string
}
\description{
This function performs the extended median test
}

