% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-CountDataTests.R
\name{fisher.test}
\alias{fisher.test}
\title{Fisher test for counts of data}
\usage{
fisher.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Fisher test
}

