% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{fisher.test}
\alias{fisher.test}
\title{Fisher test for count data}
\usage{
fisher.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
List of a htest object with pvalues, statistic and method name and a report of the test in a string
}
\description{
This function performs the Fisher test
}

