% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{freundAnsariBradley.test}
\alias{freundAnsariBradley.test}
\title{Freund Ansari Bradley test for scale}
\usage{
freundAnsariBradley.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
List of a htest object with pvalues, statistic and method name and a report of the test in a string
}
\description{
This function performs the Freund Ansari Bradley test
}

