% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{incompleteConcordance.test}
\alias{incompleteConcordance.test}
\title{Incomplete Concordance test for multiple comparisons}
\usage{
incompleteConcordance.test(matrix, lambda)
}
\arguments{
\item{matrix}{Matrix of data}

\item{lambda}{Parameter of the test}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Incomplete Concordance test
}

