% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{incompleteConcordance.test}
\alias{incompleteConcordance.test}
\title{Incomplete Concordance test for multiple comparisons}
\usage{
incompleteConcordance.test(matrix, lambda)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
List of a htest object with pvalues, statistic and method name and a report of the test in a string
}
\description{
This function performs the Incomplete Concordance test
}

