% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{jt.test}
\alias{jt.test}
\title{Jonckheere and Terpstra test for equality}
\usage{
jt.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Jonckheere and Terpstra test
}

