% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-BivariateTests.R, R/NP-Kendall.R, R/utils.R
\name{kendall.test}
\alias{kendall.test}
\alias{kendall.test}
\alias{kendall.test}
\title{Kendall test for bivariated samples}
\usage{
kendall.test(matrix)

kendall.test(matrix)

kendall.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}

\item{matrix}{Matrix of data}

\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics,
method and data name

A list with pvalues for alternative hypothesis, statistics,
method and data name

A list with pvalues for alternative hypothesis, statistics,
method and data name
}
\description{
This function performs the Kendall test

This function performs the Kendall test

This function performs the Kendall test
}
\examples{
x <- 1:10
m <- matrix(c(x, 2*x+rnorm(length(x))), ncol = 2)
kendall.test(m)
x <- 1:10
m <- matrix(c(x, 2*x+rnorm(length(x))), ncol = 2)
kendall.test(m)
x <- 1:10
m <- matrix(c(x, 2*x+rnorm(length(x))), ncol = 2)
kendall.test(m)
}
