% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{KendallQuantileTable}
\alias{KendallQuantileTable}
\title{Distribution table for aproximate version of Kendall test}
\format{An object of class \code{matrix} with 20 rows and 5 columns.}
\usage{
KendallQuantileTable
}
\description{
Distribution table for aproximate version of Kendall test
}
\keyword{datasets}

