% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{klotz.test}
\alias{klotz.test}
\title{Klotz test for scale}
\usage{
klotz.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Klotz test
}

