% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{KolmogorovTwoSampleAsymptotic}
\alias{KolmogorovTwoSampleAsymptotic}
\title{Table for KolmogorovTwoSampleAsymptotic}
\format{An object of class \code{matrix} with 12 rows and 5 columns.}
\usage{
KolmogorovTwoSampleAsymptotic
}
\description{
Table for KolmogorovTwoSampleAsymptotic
}
\keyword{datasets}

