% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-GoodnessTests.R, R/npjavatests.R
\name{ks.test}
\alias{ks.test}
\title{Kolmogorov-Smirnov test for goodness of fit}
\usage{
ks.test(matrix)

ks.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}

\item{sequence}{Secuence of data}

\item{distribution}{Distribution name to perform test}
}
\value{
A htest object with pvalues and statistics

A htest object with pvalues and statistics
}
\description{
This function performs the Kolmogorov-Smirnov test

This function performs the Kolmogorov-Smirnov test for two samples
}

