% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{ks.test}
\alias{ks.test}
\title{Kolmogorov Smirnov test for goodness of fit}
\usage{
ks.test(matrix)

ks.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}

\item{sequence}{Sequence of data}
}
\value{
List of a htest object with pvalues, statistic and method name and a report of the test in a string

List of a htest object with pvalues, statistic and method name and a report of the test in a string
}
\description{
This function performs the Kolmogorov Smirnov test

This function performs the Kolmogorov-Smirnov test for two samples
}

