% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{lilliefors.test}
\alias{lilliefors.test}
\title{Lilliefors test for goodness of fit}
\usage{
lilliefors.test(sequece)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
List of a htest object with pvalues, statistic and method name and a report of the test in a string
}
\description{
This function performs the Lilliefors test
}

