% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{mcNemar.test}
\alias{mcNemar.test}
\title{McNemar test for count data}
\usage{
mcNemar.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the McNemar test
}

