% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{mood.test}
\alias{mood.test}
\title{Mood test for scale}
\usage{
mood.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Mood test
}

