% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{multinomialEq.test}
\alias{multinomialEq.test}
\title{Multinomial equality test for count data}
\usage{
multinomialEq.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Multinomial equality test
}

