% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{NMRanksRight}
\alias{NMRanksRight}
\title{Table for NMRanksRight}
\format{An object of class \code{data.frame} with 54 rows and 3 columns.}
\usage{
NMRanksRight
}
\description{
Table for NMRanksRight
}
\keyword{datasets}

