% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{numberRuns.test}
\alias{numberRuns.test}
\title{Number of Runs test for randomness}
\usage{
numberRuns.test(sequence)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Number Of Runs test
}

