% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-RandomnessTests.R
\name{numberRunsUpDown.test}
\alias{numberRunsUpDown.test}
\title{Number of runs up and down median test for randomness}
\usage{
numberRunsUpDown.test(sequence)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Number of runs up and down  test
}

