% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{numberRunsUpDown.test}
\alias{numberRunsUpDown.test}
\title{Number of Runs Up and Down test for randomness}
\usage{
numberRunsUpDown.test(sequence)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Number of Runs Up and Down test
}

