% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{numberRunsUpDown.test}
\alias{numberRunsUpDown.test}
\title{Number of Runs Up and Down test for randomness}
\usage{
numberRunsUpDown.test(sequence)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
List of a htest object with pvalues, statistic and method name and a report of the test in a string
}
\description{
This function performs the Number of Runs Up and Down test
}

