% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{numberRunsUpDownMedian.test}
\alias{numberRunsUpDownMedian.test}
\title{Number of Runs Up and Down Median test for randomness}
\usage{
numberRunsUpDownMedian.test(sequence)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Number of Runs Up and Down Median test
}

