% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-Page.R
\name{page.test}
\alias{page.test}
\title{Page test for multiple comparisons}
\usage{
page.test(x, y = NULL)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A list with pvalues for alternative hypothesis, statistics, method and data name
}
\description{
This function performs the Page test
}
