% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-MultipleTests.R
\name{page.test}
\alias{page.test}
\title{Page test for multiple comparisons}
\usage{
page.test(matrix)

page.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}

\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics

A htest object with pvalues and statistics
}
\description{
This function performs the Page test

This function performs the Partial correlation test
}

