% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{partialCorrelation.test}
\alias{partialCorrelation.test}
\title{Partial correlation test for multiple comparisons}
\usage{
partialCorrelation.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the partial correlation test
}

