% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{partialCorrelation.test}
\alias{partialCorrelation.test}
\title{Partial correlation test for multiple comparisons}
\usage{
partialCorrelation.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
List of a htest object with pvalues, statistic and method name and a report of the test in a string
}
\description{
This function performs the partial correlation test
}

