% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{PartialCorrelationTable}
\alias{PartialCorrelationTable}
\title{Table for PartialCorrelationTable}
\format{An object of class \code{matrix} with 20 rows and 4 columns.}
\usage{
PartialCorrelationTable
}
\description{
Table for PartialCorrelationTable
}
\keyword{datasets}

