% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotPosterior}
\alias{plotPosterior}
\title{Plot of posterior distribution}
\usage{
plotPosterior(x, y, names, ...)
}
\arguments{
\item{x}{Sequence in x-axis}

\item{y}{Values of difference distribution}

\item{names}{Names of the algorithms}
}
\description{
This function plots the posterior distribution of the parameter
}

