% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSimplex}
\alias{plotSimplex}
\title{Projection of 3-simplex points}
\usage{
plotSimplex(points)
}
\arguments{
\item{points}{Simplex points to be projected}
}
\description{
This function projects 3-simplex points to a 2D triangle
}

