% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{populationQuantile.test}
\alias{populationQuantile.test}
\title{Population quantile for one sample}
\usage{
populationQuantile.test(sequence, quantile, value)
}
\value{
Report of the test in a string
}
\description{
This function performs the confidence quantile test
}

