% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{RanksVonNeumann}
\alias{RanksVonNeumann}
\title{Table for RanksVonNeumann}
\format{An object of class \code{matrix} with 41 rows and 5 columns.}
\usage{
RanksVonNeumann
}
\description{
Table for RanksVonNeumann
}
\keyword{datasets}

