% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{RunsUpDown}
\alias{RunsUpDown}
\title{Distribution table for Runs Up and Down test}
\format{An object of class \code{data.frame} with 299 rows and 3 columns.}
\usage{
RunsUpDown
}
\description{
Distribution table for Runs Up and Down test
}
\keyword{datasets}

