% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-ScaleTests.R
\name{siegelTukey.test}
\alias{siegelTukey.test}
\title{Siegel-Tukey test for scale}
\usage{
siegelTukey.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Siegel Tukey test
}

