% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{SpearmanExactTable}
\alias{SpearmanExactTable}
\title{Table for SpearmanExactTable}
\format{An object of class \code{data.frame} with 253 rows and 3 columns.}
\usage{
SpearmanExactTable
}
\description{
Table for SpearmanExactTable
}
\keyword{datasets}

