% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{SpearmanQuantileTable}
\alias{SpearmanQuantileTable}
\title{Quantile distribution table for Spearman test}
\format{An object of class \code{matrix} with 20 rows and 6 columns.}
\usage{
SpearmanQuantileTable
}
\description{
Quantile distribution table for Spearman test
}
\keyword{datasets}

