% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{sukhatme.test}
\alias{sukhatme.test}
\title{Sukhatme test for scale}
\usage{
sukhatme.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Sukhatme test
}

