% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{TotalNumberOfRuns.Left}
\alias{TotalNumberOfRuns.Left}
\title{Left distribution table for total number of runs test}
\format{An object of class \code{data.frame} with 457 rows and 4 columns.}
\usage{
TotalNumberOfRuns.Left
}
\description{
Left distribution table for total number of runs test
}
\keyword{datasets}
