% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{TotalNumberOfRuns.Right}
\alias{TotalNumberOfRuns.Right}
\title{Right distribution table for total number of runs test}
\format{An object of class \code{data.frame} with 300 rows and 4 columns.}
\usage{
TotalNumberOfRuns.Right
}
\description{
Right distribution table for total number of runs test
}
\keyword{datasets}
