% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{twoSamplesMedian.test}
\alias{twoSamplesMedian.test}
\title{Median test for two samples}
\usage{
twoSamplesMedian.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Median test
}

