% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP-RandomnessTests.R
\name{vonNeumann.test}
\alias{vonNeumann.test}
\title{Von Neumann test for randomness}
\usage{
vonNeumann.test(sequence)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Von Neumann
}

