% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{vonNeumann.test}
\alias{vonNeumann.test}
\title{Von Neumann test for randomness}
\usage{
vonNeumann.test(sequence)
}
\arguments{
\item{sequence}{Sequence of data}
}
\value{
List of a htest object with pvalues, statistic and method name and a report of the test in a string
}
\description{
This function performs the Von Neumann test
}

