% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npjavatests.R
\name{waldWolfowitz.test}
\alias{waldWolfowitz.test}
\title{Wald-Wolfowitz for two samples}
\usage{
waldWolfowitz.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
A htest object with pvalues and statistics
}
\description{
This function performs the Wald-Wolfowitz test
}

