% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{WilcoxonRanksSumTable}
\alias{WilcoxonRanksSumTable}
\title{Distribution table for Wilcoxon Ranks Sum test}
\format{An object of class \code{data.frame} with 900 rows and 4 columns.}
\usage{
WilcoxonRanksSumTable
}
\description{
Distribution table for Wilcoxon Ranks Sum test
}
\keyword{datasets}

