% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{wilcoxonRankSum.test}
\alias{wilcoxonRankSum.test}
\title{Wilcoxon Rank Sum test for location}
\usage{
wilcoxonRankSum.test(matrix)
}
\arguments{
\item{matrix}{Matrix of data}
}
\value{
List of a htest object with pvalues, statistic and method name and a report of the test in a string
}
\description{
This function performs the Wilcoxon Rank Sum test
}

