% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\docType{data}
\name{WilcoxonTable}
\alias{WilcoxonTable}
\title{Distribution table for Wilcoxon test}
\format{An object of class \code{matrix} with 504 rows and 47 columns.}
\usage{
WilcoxonTable
}
\description{
Distribution table for Wilcoxon test
}
\keyword{datasets}

